Watchlist, added content
Added content to the page, can't receive url quarries because pages aren't yet merged. Still trying to figure out how to sort the movies into watch later category.

Thinking we use session storage, use the movie id as the "name" of the variable and the movie index as the variable data. then get the variable on the watchlater page and use an if statement checking if the index of the for loop on the watchlater page matches the session storage id, then append that data.